+++ src/List.ts
@@ -16,6 +16,7 @@ export default class List<T> implements Iterator<T>, IList<T>
     private _iList: object;
     private _length: number;
     private _iCounter: number;
+    private _iForEachControlCondition: number;
 
 // ############### P R O P E R T I E S ###############
 
@@ -68,7 +69,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Adds an element at the end of the List
      * @param value Value to add
      */
-    public add(value: T): void {
+    public add(value: T): void
+    {
         if (value === undefined)
         {
             throw new Error("An undefined value cannot be added to a list");
@@ -86,7 +88,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * @param values Values as array of the type T
      */
     public addRange(values: T[]): void;
-    public addRange(values: T[] | List<T>): void {
+    public addRange(values: T[] | List<T>): void
+    {
         if (Array.isArray(values)) {
             for (let i: number = 0; i < values.length; i++) {
                 this.add(values[i]);
@@ -100,9 +103,18 @@ export default class List<T> implements Iterator<T>, IList<T>
     }
 
     /**
+     * Called in a forEach loop before a return keyword, the loop will be terminated immediately (break)
+     */
+    public break(): void
+    {
+        this._iForEachControlCondition = 1;
+    }
+
+    /**
      * Removes all elements of the List
      */
-    public clear(): void {
+    public clear(): void
+    {
         if (this._length === 0) { return; }
         else {
             this._iList = [];
@@ -112,31 +124,45 @@ export default class List<T> implements Iterator<T>, IList<T>
 
         /**
      * Check whether the List contains the specified value
-     * @param value True if the value exists, otherwise false
+     * @param value Value to check for existence
+     * @returns True if the value exists, otherwise false
      */
-    public contains(value: T): boolean {
-        if (this._length === 0) { return false; }
-        let index: number = this.indexOf(value);
-        if (index < 0) { return false; }
-        else { return true; }
-    }
+        public contains(value: T): boolean
+        {
+            if (this._length === 0) { return false; }
+            let index: number = this.indexOf(value);
+            if (index < 0) { return false; }
+            else { return true; }
+        }
+
+        /**
+        * Optional / syntactic function: Called in a forEach before a return keyword, the current iteration will be skipped (continue). It is sufficient only to call return for the same behavior
+        */
+        public continue(): void
+        {
+            this._iForEachControlCondition = 2;
+        } 
 
     /**
      * Copies the List to a new Array of the type T, from the specified starting index until the last entry of the List
      * @param startIndex Start index
+     * @returns An Array of the copied values
      */
     public copyToArray(startIndex: number): T[];
+    /**
+     * Copies the whole List to a new Array of the type T
+     * @returns An Array of the copied values
+     */
+    public copyToArray(): T[];
     /**
      * Copies the List to a new Array of the type T, from the specified starting index to the specified end index of the List
      * @param startIndex Start index
      * @param endIndex End index
+     * @returns An Array of the copied values
      */
-    /**
-     * Copies the whole List to a new Array of the type T
-     */
-    public copyToArray(): T[];    
     public copyToArray(startIndex: number, endIndex: number): T[];
-    public copyToArray(start?: number, end?: number): T[] {
+    public copyToArray(start?: number, end?: number): T[]
+    {
         if (this._length === 0) { return new Array() as T[]; }
         if (start === undefined) { start = 0; }
         if (end === undefined) { end = this._length - 1; }
@@ -145,8 +171,10 @@ export default class List<T> implements Iterator<T>, IList<T>
 
     /**
      * Removes the top element of the List and returns its value (end position / last element). undefined will be returned if the List is empty
+     * @returns The dequeued value or undefined if the List is empty
      */
-    public dequeue(): T | undefined {
+    public dequeue(): T | undefined
+    {
         if (this._length === 0) { return undefined; }
         let value: T = this._iList[this._length - 1];
         this.removeAt(this._length - 1);
@@ -156,7 +184,8 @@ export default class List<T> implements Iterator<T>, IList<T>
     /**
      * Removes all duplicates of values in the List. All duplicates after the first occurrence of each value will be removed
      */
-    public distinct(): void {
+    public distinct(): void
+    {
         if (this._length === 0) { return; }
         let newList: List<T> = new List<T>();
         for (let i = 0; i < this._length; i++) {
@@ -172,7 +201,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Inserts a new value at the top position of the List (end position / last element). This method is synonymous with add()
      * @param value Value to insert
      */
-    public enqueue(value: T): void {
+    public enqueue(value: T): void
+    {
         this.add(value);
     }
 
@@ -181,12 +211,19 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Implementation of a forEach loop
      * @param callback Callback function to process the items of the List
      */
-    public forEach(callback: IForEachInterface<T>): void {
+    public forEach(callback: IForEachInterface<T>): void
+    {
         if (this._length === 0) { return; }
         let done: boolean = false;
         let item: IteratorItem<T>;
+        this._iForEachControlCondition = 0;
         this._iCounter = 0;
-        while (done === false) {
+        while (done === false)
+        {
+            if (this.getForEachControlCondition() === 1) // break
+            {
+                return;
+            }
             item = this.next() as IteratorItem<T>;
             done = item.isLastEntry;
             callback(item.value);
@@ -196,8 +233,10 @@ export default class List<T> implements Iterator<T>, IList<T>
     /**
      * Gets the value of the List at the specified index position
      * @param index Index position (0 to n)
+     * @returns The value at the defined index position
      */
-    public get(index: number): T {
+    public get(index: number): T
+    {
         let value: T = this._iList[index];
         if (value !== undefined) {
             return value;
@@ -211,18 +250,22 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Copies the List to a new List from the specified starting index to the specified end index of the List
      * @param startIndex Start index (0 if undefined)
      * @param endIndex End index (end index if undefined)
+     * @returns A List of values at the defined index positions
      */
     public getRange(startIndex: number, endIndex: number): List<T>;
     /**
      * Copies the List to a new List from the specified starting index until the last entry of the List
      * @param startIndex Start index
+     * @returns A List of values at the defined index positions
      */
     public getRange(startIndex: number): List<T>;
     /**
      * Copies the whole List to a new List
+     * @returns A List of all values
      */
     public getRange(): List<T>;     
-    public getRange(start?: number, end?: number): List<T> {
+    public getRange(start?: number, end?: number): List<T>
+    {
         if (start === undefined) { start = 0; }
         if (end === undefined) { end = this._length - 1; }
         return this.copyToInternal(start, end, false) as List<T>;
@@ -231,8 +274,10 @@ export default class List<T> implements Iterator<T>, IList<T>
     /**
      * Gets the index of the first occurrence of the passed value
      * @param value Value to check
+     * @returns The index position of the value in the List. If not found, -1 will be returned
      */
-    public indexOf(value: T): number {
+    public indexOf(value: T): number
+    {
 
         for(let i: number = 0; i < this._length; i++)
         {
@@ -247,15 +292,19 @@ export default class List<T> implements Iterator<T>, IList<T>
     /**
      * Gets an Array of the indices of all occurrences of the passed value
      * @param value Value to check
+     * @returns An Array of indices of the specified value
      */
-    public indicesOf(value: T): number[] {
+    public indicesOf(value: T): number[]
+    {
         return this.indicesOfInternal(value, false) as number[];
     }
     /**
      * Gets a List of the indices of all occurrences of the passed value
      * @param value Value to check
+     * @returns A List of indices of the specified value
      */
-    public indicesOfAsList(value: T): List<number> {
+    public indicesOfAsList(value: T): List<number>
+    {
         return this.indicesOfInternal(value, true) as List<number>;
     }
 
@@ -264,7 +313,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * @param index Index position where to insert the value
      * @param value Value to insert
      */
-    public insertAtIndex(index: number, value: T): void {
+    public insertAtIndex(index: number, value: T): void
+    {
         this.indexCheck(index, true);// allowed 0 to length (insert after last item)
         let firstPart, secondPart: T[];
         if (index === 0) {
@@ -288,8 +338,10 @@ export default class List<T> implements Iterator<T>, IList<T>
     /**
      * Gets the index of the last occurrence of the passed value
      * @param value Value to check
+     * @returns The last index position of the specified value in the List
      */
-    public lastIndexOf(value: T): number {
+    public lastIndexOf(value: T): number
+    {
         let indices: List<number> = this.indicesOfAsList(value);
         let len: number = indices.length;
         if (len === 0) { return -1; }
@@ -300,8 +352,10 @@ export default class List<T> implements Iterator<T>, IList<T>
      /**
      * Method to get the next value of an iterator. If the last item of the List is reached, the returned object indicates that the iterations are finished. Afterwards, the method starts again at index position 0. Calling of the forEach method will also reset the position to 0. If true (boolean) is passed as value to the method, the return value will indicate that the last item is reached (break emulation)
      * @param value Optional: If true (boolean) is passed, the current result item will indicate that is is the last entry (break emulation)
+     * @returns An IteratorResult object containing a value 
      */
-    public next(value?: any): IteratorResult<T> {
+    public next(value?: any): IteratorResult<T>
+    {
         let val: any = this._iList[this._iCounter];
         let lastItem: boolean;
         if (this._iCounter < this.length - 1) {
@@ -319,6 +373,7 @@ export default class List<T> implements Iterator<T>, IList<T>
 
     /**
      * Returns the last element of a list without removing it (end of list). Returns undefined if the list is empty
+     * @return The last value of the list. Undefined will be returned if the List is empty
      */
     public peek(): T | undefined
     {
@@ -328,8 +383,10 @@ export default class List<T> implements Iterator<T>, IList<T>
 
     /**
      * Removes the bottom element of the List and returns its value (index position 0). undefined will be returned if the List is empty
+     * @return The last removed value. Undefined will be returned if the List is empty
      */
-    public pop(): T | undefined {
+    public pop(): T | undefined
+    {
         if (this._length === 0) { return undefined; }
         let value: T = this._iList[0];
         this.removeAt(0);
@@ -340,15 +397,18 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Inserts a new value at the bottom position of the List (index position 0)
      * @param value Value to insert
      */
-    public push(value: T): void {
+    public push(value: T): void
+    {
         this.insertAtIndex(0, value);
     }
 
     /**
      * Removes the passed value at the first occurrence in the List
      * @param value Value to remove
+     * @returns True if the value could be removed
      */
-    public remove(value: T): boolean {
+    public remove(value: T): boolean
+    {
         if (this._length === 0) { return false; }
         let oIndex: number = this.indexOf(value);
         if (oIndex < 0) { return false; }
@@ -362,8 +422,10 @@ export default class List<T> implements Iterator<T>, IList<T>
     /**
      * Removes the passed value at all positions in the List 
      * @param value Value to remove
+     * @returns True if the value could be removed in the whole List
      */
-    public removeAll(value: T): boolean {
+    public removeAll(value: T): boolean
+    {
         if (this._length === 0) { return false; }
         let indices: List<number> = this.indicesOfAsList(value);
         if (indices.length === 0) {
@@ -379,7 +441,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Removes the value at the defined index. All values above will be shifted one index position down (index - 1)
      * @param index Index where to remove a value
      */
-    public removeAt(index: number): void {
+    public removeAt(index: number): void
+    {
         let i: List<number> = new List<number>(index);
         this.removeAtIndices(i);
     }
@@ -394,7 +457,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * @param indices List of indices to remove
      */
     public removeAtIndices(indices: List<number>): void;
-    public removeAtIndices(indices: List<number> | number[]): void {
+    public removeAtIndices(indices: List<number> | number[]): void
+    {
         let list: List<number>;
         if (Array.isArray(indices)) {
             list = new List<number>(indices);
@@ -422,7 +486,8 @@ export default class List<T> implements Iterator<T>, IList<T>
     /**
      * Method to reverse the List
      */
-    public reverse(): void {
+    public reverse(): void
+    {
         if (this._length === 0) { return; }
         let halfLength = Math.floor(this._length / 2);
         let i1 = 0;
@@ -439,8 +504,10 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Updates a value of the List at the specified index position
      * @param index Index position (0 to n)
      * @param value New value
+     * @throws Throws an error if an undefined values was added
      */
-    public set(index: number, value: T): void {
+    public set(index: number, value: T): void
+    {
         this.indexCheck(index);
         if (value === undefined)
         {
@@ -457,6 +524,7 @@ export default class List<T> implements Iterator<T>, IList<T>
     sort(sortFunction: ISortInterFace<T>): void;
     /**
      * Sorts the List according to the default behavior (for basic / common types) or an implemented compareTo function
+     * @throws Throws an error if no suitable sorting function could be found for the type of the values
      */
     sort(): void;    
     sort(sortFunction?: ISortInterFace<T>): void 
@@ -489,7 +557,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * @param index1 Index position 1
      * @param index2 Index position 1
      */
-    public swapValues(index1: number, index2: number): void {
+    public swapValues(index1: number, index2: number): void
+    {
         this.indexCheck(index1);
         this.indexCheck(index2);
         var temp: T = new Object as T;
@@ -513,8 +582,11 @@ export default class List<T> implements Iterator<T>, IList<T>
      * @param start Start index
      * @param end End Index
      * @param toArray If true, an Array will be returned, otherwise a List
+     * @throws Throws an error if the start index was bigger than the end index
+     * @returns An Array or List of the copied values
      */
-    private copyToInternal(start: number, end: number, toArray: boolean): T[] | List<T> {
+    private copyToInternal(start: number, end: number, toArray: boolean): T[] | List<T>
+    {
         this.indexCheck(start);
         this.indexCheck(end);
         if (start > end) {
@@ -538,6 +610,15 @@ export default class List<T> implements Iterator<T>, IList<T>
         }
         return output;
     }
+
+    /**
+     * Internal function to get the state of a forEach flow control action (break or continue)
+     * @returns Returns 1 at a break condition and 2 at a continue condition (0 is default)
+     */
+    private getForEachControlCondition(): number
+    {
+        return this._iForEachControlCondition;
+    }    
 
     /**
      * Checks the validity of an index position (>= 0 < length, integer)
@@ -570,6 +651,7 @@ export default class List<T> implements Iterator<T>, IList<T>
      * Internal method to get the indices of a value in the List
      * @param value Value to check
      * @param asList If true, a List of indices will be returned, otherwise an Array
+     * @returns An Array or List of indices
      */
     private indicesOfInternal(value: T, asList: boolean): number[] | List<number> {
         let indices: List<number> = new List<number>();
@@ -592,7 +674,8 @@ export default class List<T> implements Iterator<T>, IList<T>
      * @param index2 Index position 1
      * @param tempParameter Temporary variable (Define it once outside of this method)
      */
-    private swapValuesInternal(index1: number, index2: number, tempParameter: T): void {
+    private swapValuesInternal(index1: number, index2: number, tempParameter: T): void
+    {
         tempParameter = this._iList[index1];
         this._iList[index1] = this._iList[index2];
         this._iList[index2] = tempParameter;